+++ lib/vcr/middleware/faraday.rb
@@ -77,7 +77,7 @@ module VCR
           app.call(env)
         end
 
-        def on_stubbed_request
+        def on_stubbed_by_vcr_request
           headers = env[:response_headers] ||= ::Faraday::Utils::Headers.new
           headers.update stubbed_response.headers if stubbed_response.headers
           env.update :status => stubbed_response.status.code, :body => stubbed_response.body